+++ CHANGES.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Do not automatically replace pieces when in map editor mode. It makes it harder and less predictable to edit islands
-
+* Randomizing all teams in map editor will now properly set the capitals
 
 ### Removed