+++ CHANGES.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * The default sorting is by ARtB, then by id
   * The sorting also applies to the map editor
 * Simplify instantly creating islands by retrying to shuffle the teams until it is valid
+* Add F6 shortcut to disable all piece animations
 
 ### Changed